Move TestSolanaClient to test file
